remove duplicate code from cancel-subscription-card
